Change amount to money type

+++ db/migrate/20180320195624_create_transactions.rb
@@ -2,7 +2,7 @@ class CreateTransactions < ActiveRecord::Migration[5.1]
   def change
     create_table :transactions do |t|
       t.string :coin
-      t.integer :amount
+      t.money :amount
       t.string :currency
       t.string :exchange
